day 17 part 1 (fixed some warnings in other places)

+++ BigInt.h
@@ -956,7 +956,7 @@ BigInt BigInt::operator+(const BigInt& num) const {
     result.value = "";  // the value is cleared as the digits will be appended
     short carry = 0, sum;
     // add the two values
-    for (long i = larger.size() - 1; i >= 0; i--) {
+    for (size_t i = larger.size() - 1; i >= 0; i--) {
         sum = larger[i] - '0' + smaller[i] - '0' + carry;
         result.value = std::to_string(sum % 10) + result.value;
         carry = sum / (short)10;
@@ -1013,7 +1013,7 @@ BigInt BigInt::operator-(const BigInt& num) const {
 
     result.value = "";  // the value is cleared as the digits will be appended
     short difference;
-    long i, j;
+    size_t i, j;
     // subtract the two values
     for (i = larger.size() - 1; i >= 0; i--) {
         difference = larger[i] - smaller[i];
